// src/app/layout/models/Viewer.ts
export interface Viewer {
    _id?: string;
    image?: string;
    name?: string;
    email?: string;
    profession?: string;
    interests?: [string];
    phone?: string;
    bio?: string;
    phoneno?: string;
    dob?: string;
    bname?: string;
    bcode?: string;
    accountno?: string;
    accounttitle?: string;
}